Extract building docker image to own file
